refactor: Split tree_node.py into base_node.py and named_node.py
